ENH: Add template parameter to specify the mask image type
Add 3rd template argument to the CoocurrenceTextureFeaturesImageFilter
and the RunLengthTextureFeaturesImageFilter to allow the specification
of the Mask image type. Because the mask image essentially contains
boolean values using an unsigned byte type, maybe more memory
efficient if the primary input image is a larger pixel type. Also it
is sometime a convention to only use unsigned byte pixel for mask
images.