BUG: Manually create pixel of proper length
Use the generalized NumericTraits::SetLength method to create a output
pixel type of the proper length.

The prior approach was producing a segmentation fault with the RLE
filter when used with VectorImages and an image with dimension 64x64x1
due to the neighborhood operator exceeding the size of the image in
the z-direction.

Additionally, this usage in complicated with the C++11 implementation
using a "move" assignment/constructor, where the vector may be
pointing to the same pixel in the output buffer for the duration of
the method.